Add stack-based VM implementation.
This work will allow for better interactivity, but in the meantime,
it breaks the InteractiveEvaluator.

Evaluator is much smaller now, most of the AST-based things it did are
gone, but partly replaced with the VM.Compiler class which compiles
an AST into a list of VM instructions.

A few tests & the standard library had to be updated a bit to work with
the new runtime.